// traffic-one/functions/types/api.ts
export interface CreateProjectBody {
  name: string
  organization_slug: string
  db_pass?: string
  db_region?: string
  cloud_provider?: string
  plan?: string
}

export interface CreateProjectResponse {
  id: number
  ref: string
  name: string
  status: string
  endpoint: string
  anon_key: string
  service_key: string
  organization_id: number
  organization_slug: string
  region: string
  cloud_provider: string
  is_branch_enabled: boolean
  is_physical_backups_enabled: boolean
  preview_branch_refs: string[]
  subscription_id: string | null
  inserted_at: string
  disk_volume_size_gb?: number
  infra_compute_size?: string
}

export interface ProjectDetailResponse {
  id: number
  ref: string
  name: string
  status: string
  cloud_provider: string
  region: string
  organization_id: number
  db_host: string
  connectionString: string | null
  restUrl: string
  high_availability: boolean
  is_branch_enabled: boolean
  is_physical_backups_enabled: boolean
  subscription_id: string
  inserted_at: string
  updated_at: string
}

export interface ProjectListItem {
  id: number
  ref: string
  name: string
  status: string
  region: string
  cloud_provider: string
  organization_id: number
  organization_slug: string
  is_branch_enabled: boolean
  is_physical_backups_enabled: boolean
  preview_branch_refs: string[]
  subscription_id: string | null
  inserted_at: string
}

export interface ListProjectsPaginatedResponse {
  pagination: { count: number; limit: number; offset: number }
  projects: ProjectListItem[]
}

export interface OrgProjectDatabase {
  identifier: string
  infra_compute_size: string
  region: string
  status: string
  type: string
  cloud_provider: string
}

export interface OrgProjectItem {
  ref: string
  name: string
  status: string
  region: string
  cloud_provider: string
  inserted_at: string
  is_branch: boolean
  databases: OrgProjectDatabase[]
}

export interface OrganizationProjectsResponse {
  pagination: { count: number; limit: number; offset: number }
  projects: OrgProjectItem[]
}

export interface RemoveProjectResponse {
  id: number
  ref: string
  name: string
  status: string
}